Moved food class to food module
